# rs-order/rs-processing/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Launch process batches and monitor executions",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:9043",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "REGARDS_OAUTH2": []
    }
  ],
  "paths": {
    "/api/v1/process/config/{processBusinessId}": {
      "get": {
        "tags": [
          "process-plugin-config-controller"
        ],
        "operationId": "findByBusinessId",
        "parameters": [
          {
            "name": "processBusinessId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessPluginConfigurationRightsDTO"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorStructure"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "process-plugin-config-controller"
        ],
        "operationId": "update",
        "parameters": [
          {
            "name": "processBusinessId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProcessPluginConfigurationRightsDTO"
              }
            },
            "application/json;charset=UTF-8": {
              "schema": {
                "$ref": "#/components/schemas/ProcessPluginConfigurationRightsDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessPluginConfigurationRightsDTO"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorStructure"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "process-plugin-config-controller"
        ],
        "operationId": "save",
        "parameters": [
          {
            "name": "processBusinessId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessPluginConfigurationRightsDTO"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorStructure"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/process/fordatasets": {
      "get": {
        "tags": [
          "process-plugin-config-controller"
        ],
        "operationId": "findProcessesByDatasets",
        "parameters": [
          {
            "name": "datasets",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorStructure"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/process/config/{processBusinessId}/userRole": {
      "put": {
        "tags": [
          "process-plugin-config-controller"
        ],
        "operationId": "attachRoleToProcess",
        "parameters": [
          {
            "name": "processBusinessId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "userRole",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorStructure"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/process/metadata": {
      "get": {
        "tags": [
          "process-plugin-config-controller"
        ],
        "operationId": "listAllDetectedPlugins",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PluginMetaData"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorStructure"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/process/linkprocessdataset/{dataset}": {
      "get": {
        "tags": [
          "process-plugin-config-controller"
        ],
        "operationId": "findProcessesByDataset",
        "parameters": [
          {
            "name": "dataset",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProcessLabelDTO"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorStructure"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "process-plugin-config-controller"
        ],
        "operationId": "attachDatasetToProcesses",
        "parameters": [
          {
            "name": "dataset",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "application/json;charset=UTF-8": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorStructure"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorStructure": {
        "type": "object",
        "properties": {
          "errorId": {
            "type": "string",
            "format": "uuid"
          },
          "message": {
            "type": "string"
          },
          "time": {
            "type": "string",
            "description": "ISO formatted UTC date-time",
            "format": "date-time",
            "example": "2020-12-31T00:00:00Z"
          }
        }
      },
      "IPluginParam": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "value": {
            "type": "object"
          },
          "type": {
            "type": "string",
            "enum": [
              "STRING",
              "BYTE",
              "SHORT",
              "INTEGER",
              "LONG",
              "FLOAT",
              "DOUBLE",
              "BOOLEAN",
              "MAP",
              "COLLECTION",
              "POJO",
              "PLUGIN"
            ]
          },
          "dynamic": {
            "type": "boolean"
          }
        }
      },
      "PluginConfiguration": {
        "required": [
          "label",
          "priorityOrder"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "metaData": {
            "$ref": "#/components/schemas/PluginMetaData"
          },
          "pluginId": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "businessId": {
            "pattern": "[0-9a-zA-Z_-]*",
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "priorityOrder": {
            "type": "integer",
            "format": "int32"
          },
          "active": {
            "type": "boolean"
          },
          "parameters": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IPluginParam"
            }
          },
          "iconUrl": {
            "type": "string",
            "format": "url"
          },
          "interfaceNames": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isActive": {
            "type": "boolean",
            "writeOnly": true
          },
          "pluginClassName": {
            "type": "string"
          },
          "identifiable": {
            "type": "boolean"
          }
        }
      },
      "PluginMetaData": {
        "type": "object",
        "properties": {
          "pluginClassName": {
            "type": "string"
          },
          "interfaceNames": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "author": {
            "type": "string"
          },
          "pluginId": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "markdown": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "contact": {
            "type": "string"
          },
          "owner": {
            "type": "string"
          },
          "license": {
            "type": "string"
          },
          "parameters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PluginParamDescriptor"
            }
          },
          "licence": {
            "type": "string",
            "writeOnly": true
          }
        }
      },
      "PluginParamDescriptor": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "keyLabel": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "markdown": {
            "type": "string"
          },
          "parameterizedSubTypes": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "STRING",
                "BYTE",
                "SHORT",
                "INTEGER",
                "LONG",
                "FLOAT",
                "DOUBLE",
                "BOOLEAN",
                "MAP",
                "COLLECTION",
                "POJO",
                "PLUGIN"
              ]
            }
          },
          "type": {
            "type": "string",
            "enum": [
              "STRING",
              "BYTE",
              "SHORT",
              "INTEGER",
              "LONG",
              "FLOAT",
              "DOUBLE",
              "BOOLEAN",
              "MAP",
              "COLLECTION",
              "POJO",
              "PLUGIN"
            ]
          },
          "pluginType": {
            "type": "string"
          },
          "defaultValue": {
            "type": "string"
          },
          "optional": {
            "type": "boolean"
          },
          "sensitive": {
            "type": "boolean",
            "writeOnly": true
          },
          "unconfigurable": {
            "type": "boolean"
          },
          "sensible": {
            "type": "boolean"
          }
        }
      },
      "ProcessPluginConfigurationRightsDTO": {
        "type": "object",
        "properties": {
          "pluginConfiguration": {
            "$ref": "#/components/schemas/PluginConfiguration"
          },
          "rights": {
            "$ref": "#/components/schemas/Rights"
          }
        }
      },
      "Rights": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string"
          },
          "datasets": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
      "ProcessLabelDTO": {
        "type": "object",
        "properties": {
          "processBusinessId": {
            "type": "string",
            "format": "uuid"
          },
          "label": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "REGARDS_OAUTH2": {
        "type": "oauth2",
        "flows": {
          "password": {
            "tokenUrl": "http\\://127.0.0.1/api/v1/rs-authentication/oauth/token",
            "scopes": {}
          }
        }
      }
    }
  }
}